requirements.txt
streamlit==1.28.1
tensorflow==2.19.0
numpy==2.0.2
pillow==11.3.0
opencv-python-headless==4.12.0
